+++ src/viadot/orchestration/prefect/tasks/__init__.py
@@ -9,9 +9,11 @@ from .cloud_for_customers import cloud_for_customers_to_df
 from .customer_gauge_to_df import customer_gauge_to_df
 from .databricks import df_to_databricks
 from .dbt import dbt_task
-from .dbt_orchestration import read_dbt_manifest
-from .dbt_orchestration.trigger_downstream_node import trigger_downstream_node
-from .dbt_orchestration.update_node_state import update_node_state
+from .dbt_orchestration import (
+    read_dbt_manifest,
+    trigger_downstream_node,
+    update_node_state,
+)
 from .duckdb import duckdb_query
 from .ecb import ecb_to_df
 from .entraid import entraid_to_df
