+++ tests/unit/orchestration/dbt_dynamic/test_manifest_store.py
@@ -2,14 +2,13 @@ from unittest.mock import MagicMock, patch
 
 import pytest
 
-
-pytest.importorskip("s3fs", reason="s3fs not installed")
-
 from viadot.orchestration.dbt_dynamic.manifest_store import (
     ManifestStore,
     S3ManifestStore,
 )
 
+
+pytest.importorskip("s3fs", reason="s3fs not installed")
 
 _PATH = "s3://my-bucket/dbt/manifest.json"
 _CREDS = {
@@ -50,7 +49,9 @@ class TestS3ManifestStoreRead:
         mock_s3 = MagicMock()
         mock_s3.to_dict.return_value = _MANIFEST
 
-        with patch("viadot.orchestration.dbt_dynamic.manifest_store.S3", return_value=mock_s3):
+        with patch(
+            "viadot.orchestration.dbt_dynamic.manifest_store.S3", return_value=mock_s3
+        ):
             store = ManifestStore(store_type="s3")
             result = store.read(credentials=_CREDS, path=_PATH)
 
@@ -60,7 +61,9 @@ class TestS3ManifestStoreRead:
         mock_s3 = MagicMock()
         mock_s3.to_dict.return_value = _MANIFEST
 
-        with patch("viadot.orchestration.dbt_dynamic.manifest_store.S3", return_value=mock_s3) as mock_s3_cls:
+        with patch(
+            "viadot.orchestration.dbt_dynamic.manifest_store.S3", return_value=mock_s3
+        ) as mock_s3_cls:
             store = ManifestStore(store_type="s3")
             store.read(credentials=_CREDS, path=_PATH)
 
@@ -71,7 +74,9 @@ class TestS3ManifestStoreRead:
         mock_s3 = MagicMock()
         mock_s3.to_dict.side_effect = RuntimeError("S3 unavailable")
 
-        with patch("viadot.orchestration.dbt_dynamic.manifest_store.S3", return_value=mock_s3):
+        with patch(
+            "viadot.orchestration.dbt_dynamic.manifest_store.S3", return_value=mock_s3
+        ):
             store = ManifestStore(store_type="s3")
             with pytest.raises(RuntimeError, match="S3 unavailable"):
                 store.read(credentials=_CREDS, path=_PATH)
